pyproject.toml
[tool.poetry]
name = "binance-bot"
version = "0.1.0"
description = ""
authors = ["miguelm7 <miguelmore100@gmail.com>"]
readme = "README.md"
packages = [{include = "binance_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
ta-lib = "^0.4.25"
pandas = "^1.5.3"
requests = "^2.28.2"
binance-connector = "^2.0.0"
numpy = "^1.24.2"
plotly = "^5.14.0"
nbformat = "^5.8.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
